HW3 files
In HW3, the main source file is pthread_cpu.c file. It creates two tasks from a parent task, one which finds the no. of character which occurs three times in a source file and the other which prints CPU utilization every 100 miliseconds. Both of them print all their info of operation to a common file which is passed by the main file.